Read a file and print the number of newlines.
Use filesystem module, 'fs'.
Call fs.readFileSync(/path/to/file) to return complete contents of a file and save to var buf.
Call .toString() and .split('\n') to get an array where each index is one line from the file.
Call .length -1 to get the number of lines in the file.