You can use Node's filesystem object methods to do stuff with files.
I learned how to pass a file path as an argument from the command line, and evaluate the data in some way.

First the returned data needs to be converted to a string. The string is then split by using the newline character as the delimiter. The length of the split array minus one is the number of lines in the document.