Completes 07-http-collect learnyounode exercise
Compelted both ways that the exercise offered. The bl version wasn't as hard as I thought it would be

buffer list is a storage object for collections of Node Buffers. The toString() method on bl returns a string version of the buffer and from what I understand works in a similar way to encoding does with a response object (i.e. respone.on(...))