Adds lesson 00
So, express appears to be a module that returns a funtion that we can then run, and store the results in app.
App has a variety of methods (presumably), one of which is 'get', which takes a path and a callback. This sets up the get request to the server.
The #listen method sets the app to listen on port 3000, and also presumably sends the request? I'm not sure about this yet, but it doesn't make sense to send the request unless you're ready to listen for the response.