Complete exercise 4 - 7
Ex. 4 Needed to change it to use the extname method and convert it into a module
Ex. 5 required the module from ex.4 and used a for loop to solve the problem
Ex. 6 Used the format for the .get method and the encoding to make it into a
string
Ex.7 Used the b1 npm dependency to minimize the challenge work and simply
call the b1 method to get the data. Had to call Data.toString in the console
because it would return numbers instead of letters.