02-my-first-io finished
fs.readSyncFile is necessary but you need to require 'fs' first to read
files. The file, if given, will be passed as process.argv[2], so make sure
you pass that in as the file to rad and then work on the contents.

To avoid, toString(), this would've worked:
fs.readFileSync(process.argv[2], 'utf8').split('\n').length - 1