05-make-it-modular.js
This one was tricky. The easiest way to do it was to write it as one gigantic
function, check it, then separate it out and change a few things around so that
the module wasn't actually console logging anything.

There were a TON of parameters to keep track of. Lots of using node to try and
figure things out.