Answer to good old form
To handle a POST request, use the post() method which is used the same way as get():
  app.post('/path', function(req, res){...})

A middleware is a function invoked by Express.js before your own request handler.

A middleware is added by calling use() on the application and passing the
middleware as a parameter.

To parse x-www-form-urlencoded request bodies, Express.js can use urlencoded()
middleware from the body-parser module.

    var bodyparser = require('body-parser')
    app.use(bodyparser.urlencoded({extended: false}))