Continue with learnyounode
This is the second day of learnyounode. I am having a hard time with the
recent exercises. The prompts are quite vague and dont give enough
information for someone at my level to complete the exercise.
The documentation is not easy to digest either.

However, I now see the benifit of using promises, as passing
so many callbacks when writing solutions seems pretty unstable.